requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pygame>=2.3.0
gTTS>=2.3.1
SpeechRecognition>=3.10.0